Refactor idomaar reader to swiftly handle metadata
the iterator now works without hard-coding dataset values or
schema. It is less efficient as it now uses dictionaries instead of
vectors whether they are better or not. However, this allows one to
not to worry about dataset quirks when parsing metadata.
The iterator also has a new registry that servers as a lookup table
for existing entities, so if a session has some song it just keeps
the reference of the existing entity. This also allows metadata
to be preloaded into songs and artists.
An important change is that all elements are constructed and
persisted in cascade, even if you do not use them (users,
for example). It might be a good idea to keep a blacklist or
whitelist of entity types to save later. For now it's enough.